chore: make styles setup more robust
