+++ packages/components/.storybook/preview.js
@@ -2,7 +2,7 @@ import { setCustomElementsManifest } from '@storybook/web-components';
 import 'normalize.css';
 import '../src/solid-styles.css';
 
-import { styleDefinitions } from '../src/styles/solid-styles.declaration';
+import { fetchStyleComponents } from '../src/styles/solid-styles.declaration';
 
 import { registerIconLibrary } from '../src/utilities/icon-library';
 import { storybookUtilities } from '../scripts/storybook/helper';
@@ -12,67 +12,6 @@ import { storybookUtilities } from '../scripts/storybook/helper';
  */
 
 async function loadCustomElements() {
-  function generateFakeComponentInManifest(input) {
-    const output = {
-      declarations: [
-        {
-          kind: 'class',
-          description: '',
-          name: input.styleName,
-          slots: [{ description: 'Content of the element.', name: '' }],
-          members: [],
-          attributes: [],
-          style: true,
-          summary: input.summary,
-          status: input.status,
-          since: input.since,
-          tagName: input.styleName
-        }
-      ]
-    };
-
-    for (const attribute of input.attributes) {
-      if (attribute.options) {
-        // It's a field with multiple options.
-        output.declarations[0].members.push({
-          kind: 'field',
-          name: attribute.name,
-          type: { text: `'${attribute.options.join("' | '")}'` },
-          description: attribute.description,
-          attribute: attribute.name,
-          reflects: true
-        });
-
-        output.declarations[0].attributes.push({
-          name: attribute.name,
-          type: { text: `'${attribute.options.join("' | '")}'` },
-          description: attribute.description,
-          fieldName: attribute.name
-        });
-      } else {
-        // It's a boolean field.
-        output.declarations[0].members.push({
-          kind: 'field',
-          name: attribute.name,
-          type: { text: 'boolean' },
-          default: 'false',
-          description: attribute.description,
-          attribute: attribute.name
-        });
-
-        output.declarations[0].attributes.push({
-          name: attribute.name,
-          type: { text: 'boolean' },
-          default: 'false',
-          description: attribute.description,
-          fieldName: attribute.name
-        });
-      }
-    }
-
-    return output;
-  }
-
   fetch('./custom-elements.json')
     .then(() => {
       // Use dynamic manifest generated by Vite on runtime
@@ -84,12 +23,12 @@ async function loadCustomElements() {
       return import('../dist/custom-elements.json');
     })
     .then(customElements => {
-      console.log('Custom elements manifest loaded');
+      fetchStyleComponents().then(styleComponents => {
+        customElements.modules = [...customElements?.modules, ...styleComponents];
+        setCustomElementsManifest(customElements);
+        console.log('Custom elements manifest loaded');
+      });
 
-      let overridenCustomElements = customElements;
-      styleDefinitions.map(style => overridenCustomElements?.modules?.push(generateFakeComponentInManifest(style)));
-
-      setCustomElementsManifest(overridenCustomElements);
       return;
     });
 }
